intentos de abrir camara y tomar la foto
