+++ app/src/main/java/com/example/myclientapp/cliente/ClientActivity.java
@@ -1,18 +1,32 @@
 package com.example.myclientapp.cliente;
 
+import static android.Manifest.permission.CAMERA;
+import static android.Manifest.permission.WRITE_EXTERNAL_STORAGE;
+
+
+import androidx.activity.result.ActivityResultLauncher;
+
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 
+import android.content.DialogInterface;
+import android.content.pm.PackageManager;
+import android.net.Uri;
 import android.os.Bundle;
 import android.content.Intent;
+import android.os.Environment;
+import android.provider.MediaStore;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
+import android.widget.ImageButton;
+import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -23,18 +37,21 @@ import com.example.myclientapp.bbdd.DataBase;
 import com.example.myclientapp.notas.NotasActivity;
 import com.example.myclientapp.notas.NuevaNotaFragment;
 
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
+import java.io.File;
 
 public class ClientActivity extends AppCompatActivity {
   TextView verNom, verTel, verDir, verEmail, verOtro;
   Button btnEditar, btnVolver;
-    Fragment fragmentEditaCl;
-    FragmentTransaction fragmentTransaction;
-    LinearLayout viewCliente;
-    Cliente clModelo;
-    int id;
+  ImageButton btn_camara;
+  Fragment fragmentEditaCl;
+  FragmentTransaction fragmentTransaction;
+  LinearLayout viewCliente;
+  Cliente clModelo;
+  ActivityResultLauncher<Intent> cameraLauncher;
+  ImageView imageView;
+  File photo;
+  protected String name, path;
+  int id;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -79,6 +96,7 @@ public class ClientActivity extends AppCompatActivity {
         verOtro = findViewById(R.id.tvOtro);
         btnEditar = findViewById(R.id.btn_editar);
         btnVolver = findViewById(R.id.btn_volver);
+        btn_camara = findViewById(R.id.camara);
 
         DataBase db = new DataBase(this);
         Intent intent = getIntent();
@@ -97,6 +115,12 @@ public class ClientActivity extends AppCompatActivity {
             public void onClick(View v) {
                 Intent intent = new Intent(ClientActivity.this, MainActivity.class);
                 startActivity(intent);
+            }
+        });
+        btn_camara.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                checkForPermission();
             }
         });
 
@@ -131,4 +155,96 @@ public class ClientActivity extends AppCompatActivity {
 
         viewCliente.setVisibility(View.INVISIBLE);
     }
-}
+    ///Metodos para habilitar la camara y tomar fotos
+        private boolean checkForPermission() {
+            if((checkSelfPermission(CAMERA) == PackageManager.PERMISSION_GRANTED) && (checkSelfPermission(WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED)){
+                takePicture();
+                return true;
+            }
+            if((shouldShowRequestPermissionRationale(CAMERA))||(shouldShowRequestPermissionRationale(WRITE_EXTERNAL_STORAGE))){
+                loadRecomendationDialog();
+            }else {
+                requestPermissions(new String[]{WRITE_EXTERNAL_STORAGE, CAMERA}, 100);
+            }
+            return false;
+        }
+
+        @Override
+        public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+            super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+
+            if(requestCode == 100){
+                if(grantResults.length==2 && grantResults[0]==PackageManager.PERMISSION_GRANTED && grantResults[1]==PackageManager.PERMISSION_GRANTED){
+                   takePicture();
+                }
+            }
+        }
+
+        private void loadRecomendationDialog() {
+            AlertDialog.Builder dialog = new AlertDialog.Builder(ClientActivity.this);
+            dialog.setTitle("Permisos desactivados");
+            dialog.setMessage("Debe aceptar los permisos para poder tomar fotos");
+            dialog.setPositiveButton("Aceptar", new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialogInterface, int which) {
+                    requestPermissions(new String[]{WRITE_EXTERNAL_STORAGE, CAMERA}, 100);
+                }
+            });
+            dialog.show();
+        }
+    public void takePicture(){
+        String ROOT_FOLDER = "myClientTrack/";
+        String IMAGE_ROOT = ROOT_FOLDER + "images";
+        File fileImg = new File(Environment.getExternalStorageDirectory(), IMAGE_ROOT);
+        boolean isTaken = fileImg.exists();
+
+        if(!isTaken){
+            isTaken=fileImg.mkdirs();
+        }
+        if(isTaken){
+            name= (System.currentTimeMillis()/1000)+".jpg";
+        }
+        path=Environment.getExternalStorageDirectory()+ File.separator+ IMAGE_ROOT +File.separator+name;
+
+        photo = new File(path);
+
+        // hago la llamada a la CAMARA
+
+
+         Intent takePicIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+         takePicIntent.putExtra(MediaStore.EXTRA_OUTPUT, Uri.fromFile(photo));
+         takePicIntent.putExtra("idCliente",id);
+         startActivityForResult(takePicIntent, 1);    //********* DEPRECATE********
+
+/**
+        cameraLauncher = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), new ActivityResultCallback<ActivityResult>(){
+            @Override
+            public void onActivityResult(ActivityResult result) {
+                if(result.getResultCode()==RESULT_OK){
+                    assert result.getData() != null;
+                    Bundle extras = result.getData().getExtras();
+                    assert extras != null;
+                    Bitmap imgBitmap = (Bitmap) extras.get("data");
+                    // imgView.setImageBitmap(imgBitmap);
+                    Uri uriPath = (Uri) extras.get("uriData");
+                    imageView.setImageURI(uriPath);
+                }
+
+            }
+        });
+        **/
+    }
+
+        /**
+    public void setPicture(){
+        cameraLauncher.launch(new Intent(MediaStore.ACTION_IMAGE_CAPTURE));
+        Intent intent= new Intent();
+        intent.putExtra("uriData", Uri.fromFile(photo));
+    }
+         */
+
+
+
+
+
+  }
